requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.13.5
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10
markdownify==0.13.1
numpy==1.26.3
pandas==2.3.2
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-jobspy==1.1.82
pytz==2025.2
regex==2024.11.6
requests==2.32.5
six==1.17.0
soupsieve==2.8
tls-client==1.0.1
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0